mnt: integrating comments
